team mem2 commit

+++ src/main/java/com/laundrymanagementsystem/app/service/impl/InventorySreviceImpl.java
@@ -1,5 +1,6 @@
 package com.laundrymanagementsystem.app.service.impl;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,8 +17,11 @@ import com.laundrymanagementsystem.app.dto.ApiResponseDto.ApiResponseDtoBuilder;
 import com.laundrymanagementsystem.app.dto.InventoryRequestDto;
 import com.laundrymanagementsystem.app.mapper.CustomMapper;
 import com.laundrymanagementsystem.app.model.Inventory;
+import com.laundrymanagementsystem.app.model.User;
 import com.laundrymanagementsystem.app.repository.InventoryRepository;
+import com.laundrymanagementsystem.app.repository.UserRepository;
 import com.laundrymanagementsystem.app.service.IInventoryService;
+import com.laundrymanagementsystem.app.utility.Utility;
 
 @Service
 public class InventorySreviceImpl implements IInventoryService {
@@ -26,14 +30,21 @@ public class InventorySreviceImpl implements IInventoryService {
 	InventoryRepository inventoryRepository;
 	@Autowired
 	CustomMapper customMapper;
-
+	@Autowired
+	UserRepository userRepository;
 
 	@Override
 	public void addItem(@Valid @RequestBody(required = true) InventoryRequestDto inventoryRequestDto,
 			ApiResponseDtoBuilder apiResponseDtoBuilder, HttpServletRequest request) {
-		Inventory inventory = customMapper.inventoryRequestDtoToInventory(inventoryRequestDto);
-		saveItem(inventory);
-		apiResponseDtoBuilder.withMessage("Inventory Add Sucessfully").withStatus(HttpStatus.OK).withData(inventory);
+		User sessionUser = Utility.getSessionUser(userRepository);
+		if ((sessionUser.getRole() == 2 && sessionUser != null)||(sessionUser.getRole() == 1 && sessionUser != null)) {
+			Inventory inventory = customMapper.inventoryRequestDtoToInventory(inventoryRequestDto);
+			inventory.setCreatedAt(new Date());
+			saveItem(inventory);
+			apiResponseDtoBuilder.withMessage(Constants.ITEM_ADD_SUCCESS).withStatus(HttpStatus.OK).withData(inventory);
+		} else {
+			apiResponseDtoBuilder.withMessage(Constants.UNAUTHORIZED).withStatus(HttpStatus.UNAUTHORIZED);
+		}
 	}
 
 	private void saveItem(Inventory inventory) {
@@ -42,35 +53,63 @@ public class InventorySreviceImpl implements IInventoryService {
 
 	@Override
 	public void getAllItem(ApiResponseDtoBuilder apiResponseDtoBuilder) {
-		List<Inventory> itemList = inventoryRepository.findAll();
-		apiResponseDtoBuilder.withMessage("success").withStatus(HttpStatus.OK).withData(itemList);
-		
+		User sessionUser = Utility.getSessionUser(userRepository);
+		if ((sessionUser.getRole() == 2 && sessionUser != null)||(sessionUser.getRole() == 1 && sessionUser != null)) {
+			List<Inventory> itemList = inventoryRepository.findAll();
+			apiResponseDtoBuilder.withMessage(Constants.SUCCESSFULLY).withStatus(HttpStatus.OK).withData(itemList);
+		} else {
+			apiResponseDtoBuilder.withMessage(Constants.UNAUTHORIZED).withStatus(HttpStatus.UNAUTHORIZED);
+		}
 	}
 
 	@Override
 	public void deleteItemById(long id, ApiResponseDtoBuilder apiResponseDtoBuilder) {
-		Optional<Inventory> item = inventoryRepository.findById(id);
-		if (item.isPresent()) {
-			inventoryRepository.deleteById(id);
+		User sessionUser = Utility.getSessionUser(userRepository);
+		if ((sessionUser.getRole() == 2 && sessionUser != null)||(sessionUser.getRole() == 1 && sessionUser != null)) {
+			Optional<Inventory> item = inventoryRepository.findById(id);
+			if (item.isPresent()) {
+				inventoryRepository.deleteById(id);
 
-			apiResponseDtoBuilder.withMessage("Inventory Deleted Successfully").withStatus(HttpStatus.OK);
+				apiResponseDtoBuilder.withMessage(Constants.ITEM_DELETE_SUCCESS).withStatus(HttpStatus.OK);
+			} else {
+				apiResponseDtoBuilder.withMessage("fail").withStatus(HttpStatus.NOT_FOUND);
+			}
 		} else {
-			apiResponseDtoBuilder.withMessage("fail").withStatus(HttpStatus.NOT_FOUND);
+			apiResponseDtoBuilder.withMessage(Constants.UNAUTHORIZED).withStatus(HttpStatus.UNAUTHORIZED);
 		}
-		
 	}
 
 	@Override
 	public void updateItem(@Valid Inventory inventory, ApiResponseDtoBuilder apiResponseDtoBuilder) {
-		Optional<Inventory> inventoryChaeck = inventoryRepository.findById(inventory.getId());
-		if(inventoryChaeck.isPresent()) {
-			inventoryRepository.save(inventory);
-			apiResponseDtoBuilder.withMessage("Inventory Update Successfully.").withStatus(HttpStatus.OK)
-					.withData(inventory);
-		}else {
-			apiResponseDtoBuilder.withMessage(Constants.NO_ITEM_EXISTS).withStatus(HttpStatus.OK);
-			return;
+		User sessionUser = Utility.getSessionUser(userRepository);
+		if ((sessionUser.getRole() == 2 && sessionUser != null)||(sessionUser.getRole() == 1 && sessionUser != null)) {
+			Optional<Inventory> inventoryChaeck = inventoryRepository.findById(inventory.getId());
+			if (inventoryChaeck.isPresent()) {
+				inventory.setUpdatedAt(new Date());
+				inventoryRepository.save(inventory);
+				apiResponseDtoBuilder.withMessage("Inventory Update Successfully.").withStatus(HttpStatus.OK)
+						.withData(inventory);
+			} else {
+				apiResponseDtoBuilder.withMessage(Constants.NO_ITEM_EXISTS).withStatus(HttpStatus.OK);
+				return;
+			}
+		} else {
+			apiResponseDtoBuilder.withMessage(Constants.UNAUTHORIZED).withStatus(HttpStatus.UNAUTHORIZED);
 		}
-		
+	}
+
+	@Override
+	public void getInventoryById(ApiResponseDtoBuilder apiResponseDtoBuilder, long id) {
+		User sessionUser = Utility.getSessionUser(userRepository);
+		if ((sessionUser.getRole() == 2 && sessionUser != null)||(sessionUser.getRole() == 1 && sessionUser != null)) {
+			Optional<Inventory> inventory = inventoryRepository.findById(id);
+			if (inventory.isPresent()) {
+				apiResponseDtoBuilder.withMessage("Inventory Details ..").withStatus(HttpStatus.OK).withData(inventory);
+			} else {
+				apiResponseDtoBuilder.withMessage(Constants.NO_ITEM_EXISTS).withStatus(HttpStatus.NOT_FOUND);
+			}
+		} else {
+			apiResponseDtoBuilder.withMessage(Constants.UNAUTHORIZED).withStatus(HttpStatus.UNAUTHORIZED);
+		}
 	}
 }
